+++ Manatee.Json/Schema/Keywords/ExclusiveMaximumKeyword.cs
@@ -74,6 +74,7 @@ namespace Manatee.Json.Schema
 				results.IsValid = false;
 				results.AdditionalInfo["upperBound"] = Value;
 				results.AdditionalInfo["actual"] = context.Instance;
+				results.ErrorMessage = ErrorTemplate.ResolveTokens(results.AdditionalInfo);
 			}
 
 			return results;