+++ docs_source/release-notes.md
@@ -1,5 +1,7 @@
 # 11.0.0
 
+*v11.0.0 beta 5 - Released on 21 Jul, 2019*
+
 *v11.0.0 beta 4 - Released on 19 Jul, 2019*
 
 *v11.0.0 beta 3 - Released on 18 Jul, 2019*
@@ -25,6 +27,11 @@ In order to support some new independent reference tests, some changes were made
 - `JsonSchemaOptions.OutputFormat` now has a default value of `Flag`, which only returns whether an instance is valid, without any error details.
 
 Additionally, it was pointed out on [the JSON Schema spec repo](https://github.com/json-schema-org/json-schema-spec/issues/759) that `format` is not specifically intended for strings, but can also be used to validate other types.  To address this, the `StringFormat` type has been renamed to `Format` and now accepts `JsonValue` instead of merely `string`.
+
+### Obsolete items removed
+
+- `SchemaErrorMessages` - Error messages are now customizable via the `ErrorTemplate` static property on those keywords which have it.
+- `FluentBuilderExtensions.Items(this JsonSchema, params JsonSchema[])` - The compiler could not distinguish between the single-schema usage of `items` and the array-usage which only contained a single schema.  Instead, multiple calls to `Item` should be used.  This is also more consistent with how `Definition` and `Property` work.
 
 ## Non-breaking changes